+++ src/main/java/xyz/kohara/stellarity/mixin/mob_effects/LivingEntityMixin.java
@@ -3,11 +3,11 @@ package xyz.kohara.stellarity.mixin.mob_effects;
 import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 
-import net.minecraft.server.level.ServerLevel;
+
 import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.damagesource.DamageSources;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
@@ -16,18 +16,31 @@ import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import xyz.kohara.stellarity.registry.StellarityDamageTypes;
 import xyz.kohara.stellarity.registry.StellarityMobEffects;
 import xyz.kohara.stellarity.utils.DamageUtility;
 
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
 
 //? 1.20.1 {
 
 //? } else {
-/*import net.minecraft.core.Holder;
- *///? }
+
+/*import net.minecraft.tags.EntityTypeTags;
+
+import net.minecraft.core.Holder;
+    *///? }
+
+//? > 1.21.10 {
+/*import net.minecraft.server.level.ServerLevel;
+ *///?}
 
 @Mixin(LivingEntity.class)
 public abstract class LivingEntityMixin extends Entity {
@@ -35,9 +48,19 @@ public abstract class LivingEntityMixin extends Entity {
     @Shadow
     @Final
     private Map</*? 1.20.1 { */MobEffect/*? } else { */ /*Holder<MobEffect> *//*? }*/, MobEffectInstance> activeEffects;
+
+    @Shadow
+        //? 1.20.1
+    public abstract boolean hasEffect(MobEffect par1);
+
+    //? > 1.21
+    //public abstract boolean hasEffect(Holder<MobEffect> par1);
+
+
     @Unique
     private static DamageUtility damageUtility;
 
+    @Unique
     private boolean appliedBrittle = false;
 
     public LivingEntityMixin(EntityType<?> entityType, Level level) {
@@ -49,18 +72,21 @@ public abstract class LivingEntityMixin extends Entity {
     private boolean applyBrittleEffect(/*? if > 1.21.10 >> 'Dam' *//*ServerLevel serverLevel, */DamageSource damageSource, float f, Operation<Boolean> original) {
         boolean hurt = original.call(/*? if > 1.21.10 >> 'dam' *//*serverLevel, */damageSource, f);
 
-        if (!activeEffects.containsKey(StellarityMobEffects.BRITTLE) || !hurt || appliedBrittle) return hurt;
+        if (!activeEffects.containsKey(StellarityMobEffects.BRITTLE) || !hurt || appliedBrittle || damageSource.is(StellarityDamageTypes.BRITTLE))
+            return hurt;
 
         int amplifier = activeEffects.get(StellarityMobEffects.BRITTLE).getAmplifier();
 
         var sources = this.damageSources();
+
+
         if (damageUtility == null) damageUtility = DamageUtility.builder()
-            .setDamageSource(sources.source(StellarityDamageTypes.NO_KNOCKBACK_IGNORES_IFRAMES))
-            .setApDamageSource(sources.source(StellarityDamageTypes.ARMOR_PIERCING_NO_KB))
+            .setDamageSource(sources.source(StellarityDamageTypes.BRITTLE))
+            .setApDamageSource(sources.source(StellarityDamageTypes.BRITTLE))
             .setApRatio(0.4f)
             .build();
 
-        setTicksFrozen(160);
+        setTicksFrozen(150);
 
         appliedBrittle = true;